# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inea Scents API",
        "description": "API Documentation for Inea Scents",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get user's bookings",
                "description": "Returns a list of all bookings for the authenticated user.",
                "operationId": "4535e967ef15bc76dadf1efa40192f0d",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Booking"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "description": "Creates a new booking for the authenticated user",
                "operationId": "43e6dcca842886074084b8db86af7802",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "package_id",
                                    "customer_name",
                                    "customer_email",
                                    "pax",
                                    "event_date",
                                    "venue_address",
                                    "payment_method",
                                    "consent_privacy_version"
                                ],
                                "properties": {
                                    "package_id": {
                                        "type": "integer"
                                    },
                                    "customer_name": {
                                        "type": "string"
                                    },
                                    "customer_email": {
                                        "type": "string"
                                    },
                                    "customer_phone": {
                                        "type": "string"
                                    },
                                    "pax": {
                                        "type": "integer"
                                    },
                                    "event_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "event_time": {
                                        "type": "string",
                                        "format": "time"
                                    },
                                    "venue_address": {
                                        "type": "string"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "online",
                                            "cash"
                                        ]
                                    },
                                    "scent_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "consent_privacy_version": {
                                        "type": "string",
                                        "example": "2026-10-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Booking"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Booking": {
                "title": "Booking",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "booking_reference": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "customer_name": {
                        "type": "string"
                    },
                    "customer_email": {
                        "type": "string"
                    },
                    "customer_phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "pax": {
                        "type": "integer"
                    },
                    "event_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "event_time": {
                        "type": "string",
                        "nullable": true
                    },
                    "venue_address": {
                        "type": "string"
                    },
                    "payment_method": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "checkout_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "inquiry_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "consent_privacy_version": {
                        "type": "string",
                        "nullable": true
                    },
                    "consented_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "package": {
                        "$ref": "#/components/schemas/Package"
                    },
                    "scents": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Scent"
                        }
                    }
                },
                "type": "object"
            },
...
            "Package": {
                "title": "Package",
                "description": "Package model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Romantic Getaway"
                    },
                    "description": {
                        "type": "string",
                        "example": "A nice package for couples.",
                        "nullable": true
                    },
                    "inclusions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "pax_options": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "pax_prices": {
                        "type": "object",
                        "example": {
                            "50": 4499
                        }
                    },
                    "freebies": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 199.99
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "reviews_count": {
                        "type": "integer",
                        "example": 120
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "gallery_images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "scents": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Scent"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Scent": {
                "title": "Scent",
                "description": "Scent model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Lavender"
                    },
                    "description": {
                        "type": "string",
                        "example": "A calming floral scent.",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/lavender.jpg",
                        "nullable": true
                    },
                    "is_available": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}